+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### FIXED
+- **CSD-113**
+  - Fixed issue with with bulk granule deletion where CMR records were not being removed
+
 ### Added
 
 - **CUMULUS-4842**
@@ -23,7 +27,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     migration and standard migrations are consistent and produce identical database schemas.
 
 ### Changed
-
+- **CUMULUS-4694**
+  - Add flag to enable/disable iceberg replication
 - **CUMULUS-4882**
   - Updated the triggers on the granules table to track collection updates and introduced a
     `cumulus.allow_collection_update` setting to authorize cross-collection shifts.